backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
celery>=5.3.0
redis>=5.0.0
PyPDF2>=3.0.0
litellm>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

